Adjust responsive website, add hamberger to navbar, and make scroll behavior smooth
