+++ package.json
@@ -18,6 +18,7 @@
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-icons": "^5.3.0",
+    "react-scroll": "^1.9.0",
     "react-transition-group": "^4.4.5"
   },
   "devDependencies": {
